+++ src/utils/error.ts
@@ -8,6 +8,13 @@ export class UserError extends Error {
   }
 }
 
+export const getCodeFromUnknownError = (error: unknown): string | undefined => {
+  if (error instanceof Error && 'code' in error && typeof error.code === 'string') {
+    return error.code;
+  }
+  return undefined;
+};
+
 export const getMessageFromUnknownError = (error: unknown): string => {
   let message = 'An unknown error has occurred.';
   if (error instanceof AxiosError) {
@@ -38,7 +45,9 @@ const getErrorMessageFromAxiosError = (error: AxiosError): string => {
     message = 'The service is currently unavailable. Please try again later.';
   } else if ((error.response?.data as any)?.message) {
     message = (error.response?.data as any)?.message;
-  } else if ((error.response?.data as any)?.error?.issues[0]?.message) {
+  } else if (Array.isArray((error.response?.data as any)?.error) && (error.response?.data as any).error[0]?.message) {
+    message = (error.response?.data as any).error[0].message;
+  } else if ((error.response?.data as any)?.error?.issues?.[0]?.message) {
     message = (error.response?.data as any).error.issues[0].message;
   } else if (error.response?.data && typeof error.response?.data === 'string') {
     message = error.response.data;